+++ samples/sampler/stories/qa/VirtualList.js
@@ -3,14 +3,14 @@ import Item from '@enact/moonstone/Item';
 import {ActivityPanels, Panel, Header} from '@enact/moonstone/Panels';
 import Scroller from '@enact/moonstone/Scroller';
 import SwitchItem from '@enact/moonstone/SwitchItem';
+import {ScrollableBase as UiScrollableBase} from '@enact/moonstone/UiScrollable';
+import {VirtualListBase as UiVirtualListBase} from '@enact/moonstone/UiVirtualList';
 import VirtualList, {VirtualListBase} from '@enact/moonstone/VirtualList';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, number, select} from '@enact/storybook-utils/addons/controls';
 import ri from '@enact/ui/resolution';
-import {ScrollableBase as UiScrollableBase} from '@enact/ui/Scrollable';
-import {VirtualListBase as UiVirtualListBase} from '@enact/ui/VirtualList';
 import PropTypes from 'prop-types';
 import {useCallback, useEffect, useRef, useState} from 'react';